MAT-6856 refactored code
