+++ src/terminologySectionPanel/__mocks__/MockMeasureStoreCql.tsx
